refactor(useModal): useModal 훅 생성
