+++ src/constants/messages.ts
@@ -1,5 +1,6 @@
 export const ERROR_MESSAGES = {
   failGetCartItems: "장바구니 정보를 불러오는데 실패했습니다. 다시 시도해 주세요.",
+  failGetCartItemsQuantity: "장바구니의 개수를 불러오는데 실패했습니다. 다시 시도해 주세요.",
   failPostCartItem: "장바구니에 상품을 추가하지 못했습니다. 다시 시도해 주세요.",
   failDeleteCartItem: "장바구니에 상품을 삭제하지 못했습니다. 다시 시도해 주세요.",
   failGetProducts: "상품 목록을 가져오는데 실패했습니다. 다시 시도해 주세요.",
