+++ src/hooks/useModal.ts
@@ -1,0 +1,26 @@
+import { useBodyScrollLock } from "@/hooks/useBodyScrollLock";
+import { useState } from "react";
+
+const useModal = () => {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const { lockScroll, openScroll } = useBodyScrollLock();
+
+  const onOpenModal = () => {
+    setIsModalOpen(true);
+    lockScroll();
+  };
+
+  const onCloseModal = () => {
+    setIsModalOpen(false);
+    openScroll();
+  };
+
+  return {
+    onOpenModal,
+    onCloseModal,
+    isModalOpen,
+  };
+};
+
+export default useModal;
